+++ src/engine/ecs/components/Body.ts
@@ -1,0 +1,8 @@
+import { f32 } from 'bitecs/serialization';
+
+export const Body = {
+	offsetX: f32([]),
+	offsetY: f32([]),
+	width: f32([]),
+	height: f32([]),
+};