+++ src/engine/ecs/components/BodyTouching.ts
@@ -1,0 +1,8 @@
+import { u8 } from 'bitecs/serialization';
+
+export const BodyTouching = {
+	left: u8([]),
+	right: u8([]),
+	top: u8([]),
+	bottom: u8([]),
+};