+++ src/engine/ecs/components/SpriteFrame.ts
@@ -1,0 +1,8 @@
+import { u16 } from 'bitecs/serialization';
+
+export const SpriteFrame = {
+	x: u16([]),
+	y: u16([]),
+	width: u16([]),
+	height: u16([]),
+};