+++ src/engine/ecs/components/WorldBoundsBody.ts
@@ -1,0 +1,6 @@
+import { u8 } from 'bitecs/serialization';
+
+export const WorldBoundsBody = {
+	bounceX: u8([]),
+	bounceY: u8([]),
+};