+++ src/game/TemplateExample.ts
@@ -1,4 +1,3 @@
-import { createDomAssetCollections } from 'engine/assets/createDomAssetCollections';
 import { Game } from 'engine/core/Game';
 import { controls } from 'game/config/controls';
 import { SCREEN_HEIGHT, SCREEN_WIDTH } from 'game/constants/game';
@@ -10,10 +9,10 @@ export class TemplateExample extends Game {
 	constructor() {
 		super('body', SCREEN_WIDTH, SCREEN_HEIGHT, {
 			controls,
-			assets: createDomAssetCollections({
-				images: ['img#logo'],
-				audio: ['audio#bgm'],
-			}),
+			assetManifest: {
+				images: [new URL('../../images/logo-atlas.svg', import.meta.url).href],
+				audio: [new URL('../../sound/area12.ogg', import.meta.url).href],
+			},
 		});
 	}
 }