+++ src/game/constants/assets.ts
@@ -5,3 +5,7 @@ export const DemoImageAsset = {
 export const DemoAudioAsset = {
 	BGM: 0,
 } as const;
+
+export const DemoAnimationAsset = {
+	LOGO_IDLE: 0,
+} as const;