+++ src/game/constants/game.ts
@@ -10,3 +10,5 @@ export const SCREEN_WIDTH = 640;
 export const SCREEN_HEIGHT = 360;
 export const WORLD_WIDTH = SCREEN_WIDTH * 2.5;
 export const WORLD_HEIGHT = SCREEN_HEIGHT * 2;
+export const PORTAL_WIDTH = 120;
+export const PORTAL_HEIGHT = 120;